Refactor date util in the unit tests
